requirements.txt
Flask
gunicorn
pyyaml